requirements.txt
praw
pandas
nltk
textblob
python-dotenv
